eskiBot/requirements.txt
requests==2.31.0
urllib3==2.0.7
flask==2.3.3
flask-socketio==5.3.6
python-socketio==5.8.0
eventlet==0.33.3
sqlalchemy==2.0.23
alembic==1.12.1
mysql-connector-python==8.2.0
